+++ src/Exceptions/ExceptionInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Qless\Exceptions;
+
+/**
+ * Qless\Exceptions\ExceptionInterface
+ *
+ * @package Qless\Exceptions
+ */
+interface ExceptionInterface extends \Throwable
+{
+}